requirements.txt
certifi==2021.5.30
babel==2.9.1
flask-babel==2.0.0
flask==2.0.1
jinja2==3.0.1
lxml==4.6.3
pygments==2.10.0
python-dateutil==2.8.2
pyyaml==5.4.1
httpx[http2]==0.19.0
Brotli==1.0.9
uvloop==0.16.0; python_version >= '3.7'
uvloop==0.14.0; python_version < '3.7'
httpx-socks[asyncio]==0.4.1
langdetect==1.0.9
setproctitle==1.2.2
guppy3==3.1.1
